// packages/cli-server/generated/cli.zod.ts
export const CliAuthDeviceAuthorizationPollBody = zod.object({
  "deviceCode": zod.string().min(1)
})


export const cliAuthDeviceAuthorizationPollResponseDataOnePollAfterMsExclusiveMin = 0;
export const cliAuthDeviceAuthorizationPollResponseDataOnePollAfterMsMax = 9007199254740991;


export const cliAuthDeviceAuthorizationPollResponseDataTwoUserEmailRegExp = new RegExp('^(?!\\.)(?!.\*\\.\\.)([A-Za-z0-9_\'+\\-\\.]\*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]\*\\.)+[A-Za-z]{2,}$');

export const cliAuthDeviceAuthorizationPollResponseDataTwoIssuedAtRegExp = new RegExp('^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$');
export const cliAuthDeviceAuthorizationPollResponseDataTwoExpiresAtRegExp = new RegExp('^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$');



export const CliAuthDeviceAuthorizationPollResponse = zod.object({
  "requestId": zod.string().min(1),
  "data": zod.union([zod.object({
  "state": zod.enum(['pending']),
  "pollAfterMs": zod.number().gt(cliAuthDeviceAuthorizationPollResponseDataOnePollAfterMsExclusiveMin).max(cliAuthDeviceAuthorizationPollResponseDataOnePollAfterMsMax)
}),zod.object({
  "state": zod.enum(['authorized']),
  "accessToken": zod.string().min(1),
  "authMode": zod.enum(['browser_session', 'bearer_token']),
  "user": zod.object({
  "id": zod.string(),
  "email": zod.string().email().regex(cliAuthDeviceAuthorizationPollResponseDataTwoUserEmailRegExp),
  "displayName": zod.string()
}),
  "activeOrgSlug": zod.string().min(1).nullable(),
  "issuedAt": zod.string().datetime({}).regex(cliAuthDeviceAuthorizationPollResponseDataTwoIssuedAtRegExp).nullable(),
  "expiresAt": zod.string().datetime({}).regex(cliAuthDeviceAuthorizationPollResponseDataTwoExpiresAtRegExp).nullable()
})]),
  "warnings": zod.array(zod.string().min(1))
})